# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealthAI Coach API",
        "description": "Documentation de l'API HealthAI Coach, gérée avec Lomkit REST API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Se connecter et récupérer un token Sanctum",
                "operationId": "f72526a176d358410e0862d5c20c6641",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie, token retourné"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Se déconnecter (Révocation du token actuel)",
                "operationId": "3599bb21fa2c90996e7f13e2dfbbaabf",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/exercises/search": {
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Rechercher et filtrer des exercises",
                "operationId": "4c34914b8e15c867009f674f5bbbade7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/exercises/mutate": {
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Créer ou mettre à jour un/des exercises",
                "operationId": "c8864571b405af2786d84b32031fe121",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mutate": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/exercises/destroy": {
            "delete": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Supprimer un/des exercises",
                "operationId": "17a3b7d3649606d8012249ea87aa9045",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/exercises/restore": {
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Restaurer un/des exercices supprimés",
                "operationId": "2d99e40ab3b24d702b12c5b9d4104478",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/foods/search": {
            "post": {
                "tags": [
                    "Foods"
                ],
                "summary": "Rechercher et filtrer des aliments",
                "operationId": "826443feb332b916538e57b6cb109266",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/foods/mutate": {
            "post": {
                "tags": [
                    "Foods"
                ],
                "summary": "Créer ou mettre à jour un/des aliments",
                "operationId": "b6ac0509f151528051c8977b910503e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mutate": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/foods/destroy": {
            "delete": {
                "tags": [
                    "Foods"
                ],
                "summary": "Supprimer un/des aliments",
                "operationId": "774de2e13d8584981fc99145579b3f84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/foods/restore": {
            "post": {
                "tags": [
                    "Foods"
                ],
                "summary": "Restaurer un/des aliments supprimés",
                "operationId": "0fb5838cd50ec94e40d30a4f64b66178",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum ici",
                "scheme": "bearer"
            }
        }
    },
...
}